requirements-test.txt
mss>=9.0
pyautogui>=0.9
PyGetWindow>=0.0.9
Pillow>=10.0
